requirements.txt
fastapi
uvicorn
crewai
python-dotenv
litellm
python-telegram-bot